requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.54.0
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-jalali-date==1.0.2
django-js-asset==2.1.0
django-mathfilters==1.0.0
django-phonenumber-field==7.1.0
django-rosetta==0.9.9
idna==3.4
jdatetime==4.1.1
oauthlib==3.2.2
phonenumbers==8.13.17
Pillow==10.0.0
polib==1.2.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
